TTSHOP.Data/Repositories/ProductCategoryRepository.cs
using System.Collections;
using System.Linq;
using TTSHOP.Data.Infrastructure;
using TTSHOP.Model.Models;

namespace TTSHOP.Data.Repositories
{
    public interface IProductCategoryRepository : IRepository<ProductCategory>
    {
        IEnumerable GetByAlias(string alias);
    }

    public class ProductCategoryRepository : RepositoryBase<ProductCategory>, IProductCategoryRepository
    {
        public ProductCategoryRepository(IDbFactory dbFactory) : base(dbFactory)
        {
        }

        IEnumerable IProductCategoryRepository.GetByAlias(string alias)
        {
            return this.DbContext.ProductCategories.Where(x => x.Alias == alias);
        }
    }
}